API/requirements.txt
Flask>=2.0
google-cloud-pubsub>=2.0.0
nba_api>=1.1.10
requests>=2.20.0